# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reviews/user/{userDid}/reviews": {
      "get": {
        "operationId": "ReviewsController_getUserReviews",
        "parameters": [
          {
            "name": "userDid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to return (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return per page",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedReviewsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get paginated reviews for a user",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/media": {
      "get": {
        "operationId": "ReviewsController_getMediaReviews",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to return (1-based)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return per page",
            "schema": {
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "mediaType",
            "required": true,
            "in": "query",
            "description": "Media type",
            "schema": {
              "type": "string",
              "enum": [
                "movie",
                "show",
                "season",
                "episode"
              ]
            }
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "query",
            "description": "TMDB movie ID or show ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seasonNumber",
            "required": false,
            "in": "query",
            "description": "Season number for season/episode items",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "episodeNumber",
            "required": false,
            "in": "query",
            "description": "Episode number for episode items",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pinnedReviewId",
            "required": false,
            "in": "query",
            "description": "Guarantee this review id is included in the response even if community ordering would push it past the first page (used by deep links).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaReviewsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public reviews for a media item",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/canonical/{handle}/{rkey}": {
      "get": {
        "operationId": "ReviewsController_getCanonicalReview",
        "parameters": [
          {
            "name": "handle",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rkey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Canonical review resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanonicalReviewResponseDto"
                }
              }
            }
          }
        },
        "summary": "Resolve the canonical public review page for {handle}/{rkey}",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/publications/mine": {
      "get": {
        "operationId": "ReviewsController_listMyPublications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Publications listed from the user's PDS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyPublicationsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the requesting user's own site.standard.publication records (picker source)",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{reviewId}": {
      "get": {
        "operationId": "ReviewsController_getReview",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a single review by id",
        "tags": [
          "reviews"
        ]
      },
      "patch": {
        "operationId": "ReviewsController_updateReview",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a review",
        "tags": [
          "reviews"
        ]
      },
      "delete": {
        "operationId": "ReviewsController_deleteReview",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews": {
      "post": {
        "operationId": "ReviewsController_createReview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateReviewResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{reviewId}/bluesky-post": {
      "post": {
        "operationId": "ReviewsController_retryBlueskyCrossPost",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bluesky Cross-post result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueskyCrossPostResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retry a Review's one-time Bluesky Cross-post",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{reviewId}/like": {
      "post": {
        "operationId": "ReviewsController_likeReview",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review liked"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Like a review",
        "tags": [
          "reviews"
        ]
      },
      "delete": {
        "operationId": "ReviewsController_unlikeReview",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review unliked"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unlike a review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{reviewId}/likes": {
      "get": {
        "operationId": "ReviewsController_getReviewLikes",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Likes retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewLikesResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get likes for a review",
        "tags": [
          "reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "Opnshelf API",
    "description": "Personal media tracker powered by AT Protocol",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "PaginatedReviewsResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total count of items across all pages"
          },
          "page": {
            "type": "number",
            "description": "Current page number after server-side clamping"
          },
          "pageSize": {
            "type": "number",
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of available pages"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether a next page exists"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Whether a previous page exists"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserReviewDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage",
          "items"
        ]
      },
...
      "MediaReviewsResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total count of items across all pages"
          },
          "page": {
            "type": "number",
            "description": "Current page number after server-side clamping"
          },
          "pageSize": {
            "type": "number",
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of available pages"
          },
          "hasNextPage": {
            "type": "boolean",
            "description": "Whether a next page exists"
          },
          "hasPreviousPage": {
            "type": "boolean",
            "description": "Whether a previous page exists"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaReviewItemDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "totalPages",
          "hasNextPage",
          "hasPreviousPage",
          "items"
        ]
      },
      "CanonicalReviewAuthorDto": {
        "type": "object",
        "properties": {
          "did": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "required": [
          "did",
          "handle"
        ]
      },
      "CanonicalReviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rkey": {
            "type": "string",
            "description": "AT record key of the review"
          },
          "reviewTitle": {
            "type": "string"
          },
          "markdown": {
            "type": "string",
            "description": "Review body as markdown source"
          },
          "spoiler": {
            "type": "boolean",
            "description": "Author-declared Spoiler Flag: the body contains spoilers"
          },
          "description": {
            "type": "string",
            "description": "Short plaintext excerpt"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "movie",
              "show",
              "season",
              "episode"
            ]
          },
          "mediaId": {
            "type": "string"
          },
          "seasonNumber": {
            "type": "number"
          },
          "episodeNumber": {
            "type": "number"
          },
          "mediaTitle": {
            "type": "string",
            "description": "Title of the movie or show that mediaId identifies. Never composite — this is what URL slugs are built from (ADR 0023)."
          },
          "mediaLabel": {
            "type": "string",
            "description": "Human-readable label for the media, composite for a season or episode. Display only — never build a URL from it."
          },
          "posterPath": {
            "type": "string",
            "description": "Poster path of the underlying media item (the review cover)"
          },
          "author": {
            "$ref": "#/components/schemas/CanonicalReviewAuthorDto"
          },
          "canonicalUrl": {
            "type": "string",
            "description": "Absolute canonical URL on the public site (opnshelf.xyz), never the PDS host"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "rkey",
          "reviewTitle",
          "markdown",
          "spoiler",
          "mediaType",
          "mediaId",
          "author",
          "canonicalUrl",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "MyPublicationsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyPublicationDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "ReviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rkey": {
            "type": "string"
          },
          "reviewTitle": {
            "type": "string"
          },
          "markdown": {
            "type": "string",
            "description": "Review body as markdown source"
          },
          "spoiler": {
            "type": "boolean",
            "description": "Author-declared Spoiler Flag: the body contains spoilers"
          },
          "blogDocumentUri": {
            "type": "string",
            "nullable": true,
            "description": "AT-URI of the standard.site blog-mirror document, or null when not mirrored"
          },
          "mirrorToBlog": {
            "type": "boolean",
            "description": "Whether this review is mirrored to the author's blog"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "movie",
              "show",
              "season",
              "episode"
            ]
          },
          "mediaId": {
            "type": "string"
          },
          "seasonNumber": {
            "type": "number"
          },
          "episodeNumber": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "rkey",
          "reviewTitle",
          "markdown",
          "spoiler",
          "mirrorToBlog",
          "mediaType",
          "mediaId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "mediaType": {
            "type": "string",
            "description": "Media type",
            "enum": [
              "movie",
              "show",
              "season",
              "episode"
            ]
          },
          "mediaId": {
            "type": "string",
            "description": "TMDB movie ID or show ID"
          },
          "seasonNumber": {
            "type": "number",
            "description": "Season number for season/episode items"
          },
          "episodeNumber": {
            "type": "number",
            "description": "Episode number for episode items"
          },
          "title": {
            "type": "string",
            "description": "Review title (required)",
            "maxLength": 300
          },
          "markdown": {
            "type": "string",
            "description": "Review body as markdown source",
            "maxLength": 20000
          },
          "spoiler": {
            "type": "boolean",
            "description": "Author-declared Spoiler Flag: the body contains spoilers. The title stays visible everywhere and must remain spoiler-free."
          },
          "mirrorToBlog": {
            "type": "boolean",
            "description": "Whether to mirror this review to the author's blog (when one is configured). Defaults to true."
          },
          "postToBluesky": {
            "type": "boolean",
            "description": "Whether to create a one-time Bluesky post announcing this new review. Defaults to false.",
            "default": false
          }
        },
        "required": [
          "mediaType",
          "mediaId",
          "title",
          "markdown"
        ]
      },
      "BlueskyCrossPostResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "not_requested",
              "posted",
              "failed"
            ]
          },
          "uri": {
            "type": "string",
            "description": "AT-URI of the Bluesky post when it was created"
          },
          "url": {
            "type": "string",
            "description": "Public bsky.app URL for the post"
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateReviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rkey": {
            "type": "string"
          },
          "reviewTitle": {
            "type": "string"
          },
          "markdown": {
            "type": "string",
            "description": "Review body as markdown source"
          },
          "spoiler": {
            "type": "boolean",
            "description": "Author-declared Spoiler Flag: the body contains spoilers"
          },
          "blogDocumentUri": {
            "type": "string",
            "nullable": true,
            "description": "AT-URI of the standard.site blog-mirror document, or null when not mirrored"
          },
          "mirrorToBlog": {
            "type": "boolean",
            "description": "Whether this review is mirrored to the author's blog"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "movie",
              "show",
              "season",
              "episode"
            ]
          },
          "mediaId": {
            "type": "string"
          },
          "seasonNumber": {
            "type": "number"
          },
          "episodeNumber": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "blueskyCrossPost": {
            "$ref": "#/components/schemas/BlueskyCrossPostResultDto"
          }
        },
        "required": [
          "id",
          "rkey",
          "reviewTitle",
          "markdown",
          "spoiler",
          "mirrorToBlog",
          "mediaType",
          "mediaId",
          "createdAt",
          "updatedAt",
          "blueskyCrossPost"
        ]
      },
      "UpdateReviewDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Review title",
            "maxLength": 300
          },
          "markdown": {
            "type": "string",
            "description": "Review body as markdown source",
            "maxLength": 20000
          },
          "spoiler": {
            "type": "boolean",
            "description": "Author-declared Spoiler Flag: the body contains spoilers. The title stays visible everywhere and must remain spoiler-free."
          },
          "mirrorToBlog": {
            "type": "boolean",
            "description": "Whether to mirror this review to the author's blog (when one is configured)."
          }
        }
      },
...
    }
  }
}