# backend/src/reviews/reviews.controller.ts
import {
	Body,
	Controller,
	Delete,
	Get,
	NotFoundException,
	Param,
	Patch,
	Post,
	Query,
	Req,
	UseGuards,
} from "@nestjs/common";
import {
	ApiBearerAuth,
	ApiOkResponse,
	ApiOperation,
	ApiTags,
	ApiUnauthorizedResponse,
} from "@nestjs/swagger";
import { AuthGuard } from "../auth/auth.guard";
import { OptionalAuthGuard } from "../auth/optional-auth.guard";
import type { AuthenticatedRequest } from "../auth/types";
import {
	CanonicalReviewResponseDto,
	CreateReviewDto,
	CreateReviewResponseDto,
	BlueskyCrossPostResultDto,
	MediaReviewsQueryDto,
	MediaReviewsResponseDto,
	MyPublicationsResponseDto,
	PaginatedReviewsQueryDto,
	PaginatedReviewsResponseDto,
	ReviewLikesResponseDto,
	ReviewResponseDto,
	UpdateReviewDto,
} from "./dto/review.dto";
import { ReviewsService, type ATSession } from "./reviews.service";

@ApiTags("reviews")
@Controller("reviews")
export class ReviewsController {
	constructor(private readonly reviewsService: ReviewsService) {}

	@Get("user/:userDid/reviews")
	@UseGuards(OptionalAuthGuard)
	@ApiOperation({ summary: "Get paginated reviews for a user" })
	@ApiOkResponse({
		description: "Reviews retrieved",
		type: PaginatedReviewsResponseDto,
	})
	async getUserReviews(
		@Param("userDid") userDid: string,
		@Query() query: PaginatedReviewsQueryDto,
		@Req() req: AuthenticatedRequest,
	): Promise<PaginatedReviewsResponseDto> {
		const result = await this.reviewsService.getUserReviews(
			userDid,
			query.page,
			query.pageSize,
			req.user?.did,
		);

		return {
			...result.pagination,
			items: result.items.map((review) => ({
				id: review.id,
				rkey: review.rkey,
				reviewTitle: review.title,
				markdown: review.markdown,
				spoiler: review.spoiler,
				description: review.description ?? undefined,
				mediaType: review.mediaType,
				mediaId: review.mediaId,
				seasonNumber: review.seasonNumber || undefined,
				episodeNumber: review.episodeNumber || undefined,
				// `mediaLabel` is for display, `mediaTitle` is the slug source.
				mediaLabel: review.mediaLabel ?? undefined,
				mediaTitle: review.mediaTitle ?? undefined,
				posterPath: review.posterPath ?? undefined,
				likeCount: review.likeCount,
				hasLiked: review.hasLiked,
				authorRating: review.authorRating ?? undefined,
				createdAt: review.createdAt.toISOString(),
				updatedAt: review.updatedAt.toISOString(),
			})),
		};
	}

	@Get("media")
	@UseGuards(OptionalAuthGuard)
	@ApiOperation({ summary: "Get public reviews for a media item" })
	@ApiOkResponse({
		description: "Reviews retrieved",
		type: MediaReviewsResponseDto,
	})
	async getMediaReviews(
		@Query() query: MediaReviewsQueryDto,
		@Req() req: AuthenticatedRequest,
	): Promise<MediaReviewsResponseDto> {
		const result = await this.reviewsService.getMediaReviews(
			query,
			req.user?.did,
		);

		return {
			...result.pagination,
			items: result.items.map((review) => ({
				id: review.id,
				rkey: review.rkey,
				reviewTitle: review.title,
				markdown: review.markdown,
				spoiler: review.spoiler,
				description: review.description ?? undefined,
				// Canonical public review page (ADR-0013). Relative path against the
				// public site (opnshelf.xyz) — NEVER the PDS host.
				reviewUrl: `/reviews/${review.user.handle}/${review.rkey}`,
				posterPath: review.posterPath ?? undefined,
				userDid: review.user.did,
				userHandle: review.user.handle,
				userDisplayName: review.user.displayName ?? undefined,
				userAvatar: review.user.avatar ?? undefined,
				likeCount: review.likeCount,
				hasLiked: review.hasLiked,
				mirrorToBlog: review.mirrorToBlog,
				authorRating: review.authorRating ?? undefined,
				createdAt: review.createdAt.toISOString(),
				updatedAt: review.updatedAt.toISOString(),
			})),
		};
	}

	@Get("canonical/:handle/:rkey")
	@ApiOperation({
		summary: "Resolve the canonical public review page for {handle}/{rkey}",
	})
	@ApiOkResponse({
		description: "Canonical review resolved",
		type: CanonicalReviewResponseDto,
	})
	async getCanonicalReview(
		@Param("handle") handle: string,
		@Param("rkey") rkey: string,
	): Promise<CanonicalReviewResponseDto> {
		const review = await this.reviewsService.getCanonicalReview(handle, rkey);
		return {
			id: review.id,
			rkey: review.rkey,
			reviewTitle: review.title,
			markdown: review.markdown,
			spoiler: review.spoiler,
			description: review.description ?? undefined,
			mediaType: review.mediaType,
			mediaId: review.mediaId,
			seasonNumber: review.seasonNumber || undefined,
			episodeNumber: review.episodeNumber || undefined,
			mediaTitle: review.mediaTitle ?? undefined,
			mediaLabel: review.mediaLabel ?? undefined,
			posterPath: review.posterPath ?? undefined,
			author: {
				did: review.author.did,
				handle: review.author.handle,
				displayName: review.author.displayName ?? undefined,
				avatar: review.author.avatar ?? undefined,
			},
			canonicalUrl: review.canonicalUrl,
			createdAt: review.createdAt.toISOString(),
			updatedAt: review.updatedAt.toISOString(),
		};
	}

	@Get("publications/mine")
	@UseGuards(AuthGuard)
	@ApiBearerAuth()
	@ApiOperation({
		summary:
			"List the requesting user's own site.standard.publication records (picker source)",
	})
	@ApiOkResponse({
		description: "Publications listed from the user's PDS",
		type: MyPublicationsResponseDto,
	})
	@ApiUnauthorizedResponse({ description: "Not authenticated" })
	async listMyPublications(
		@Req() req: AuthenticatedRequest,
	): Promise<MyPublicationsResponseDto> {
		const items = await this.reviewsService.listMyPublications(
			req.user.did,
			req.user.session as ATSession,
		);
		return { items };
	}

	@Get(":reviewId")
	@ApiOperation({ summary: "Get a single review by id" })
	@ApiOkResponse({ description: "Review retrieved", type: ReviewResponseDto })
	async getReview(
		@Param("reviewId") reviewId: string,
	): Promise<ReviewResponseDto> {
		const review = await this.reviewsService.getReview(reviewId);
		if (!review) {
			throw new NotFoundException("Review not found");
		}
		return this.toReviewResponse(review);
	}

	@Post()
	@UseGuards(AuthGuard)
	@ApiBearerAuth()
	@ApiOperation({ summary: "Create a review" })
	@ApiOkResponse({
		description: "Review created",
		type: CreateReviewResponseDto,
	})
	@ApiUnauthorizedResponse({ description: "Not authenticated" })
	async createReview(
		@Req() req: AuthenticatedRequest,
		@Body() dto: CreateReviewDto,
	): Promise<CreateReviewResponseDto> {
		const review = await this.reviewsService.createReview(
			req.user.did,
			req.user.session as ATSession,
			dto,
		);
		return {
			...this.toReviewResponse(review),
			blueskyCrossPost: review.blueskyCrossPost,
		};
	}

	@Post(":reviewId/bluesky-post")
	@UseGuards(AuthGuard)
	@ApiBearerAuth()
	@ApiOperation({ summary: "Retry a Review's one-time Bluesky Cross-post" })
	@ApiOkResponse({
		description: "Bluesky Cross-post result",
		type: BlueskyCrossPostResultDto,
	})
	@ApiUnauthorizedResponse({ description: "Not authenticated" })
	async retryBlueskyCrossPost(
		@Param("reviewId") reviewId: string,
		@Req() req: AuthenticatedRequest,
	): Promise<BlueskyCrossPostResultDto> {
		return this.reviewsService.retryBlueskyCrossPost(
			req.user.did,
			req.user.session as ATSession,
			reviewId,
		);
	}

	@Patch(":reviewId")
	@UseGuards(AuthGuard)
	@ApiBearerAuth()
	@ApiOperation({ summary: "Update a review" })
	@ApiOkResponse({
		description: "Review updated",
		type: ReviewResponseDto,
	})
	@ApiUnauthorizedResponse({ description: "Not authenticated" })
	async updateReview(
		@Param("reviewId") reviewId: string,
		@Req() req: AuthenticatedRequest,
		@Body() dto: UpdateReviewDto,
	): Promise<ReviewResponseDto> {
		const review = await this.reviewsService.updateReview(
			req.user.did,
			req.user.session as ATSession,
			reviewId,
			dto,
		);
		return this.toReviewResponse(review);
	}

	@Delete(":reviewId")
	@UseGuards(AuthGuard)
	@ApiBearerAuth()
	@ApiOperation({ summary: "Delete a review" })
	@ApiOkResponse({ description: "Review deleted" })
	@ApiUnauthorizedResponse({ description: "Not authenticated" })
	async deleteReview(
		@Param("reviewId") reviewId: string,
		@Req() req: AuthenticatedRequest,
	): Promise<{ success: boolean }> {
		await this.reviewsService.deleteReview(
			req.user.did,
			req.user.session as ATSession,
			reviewId,
		);
		return { success: true };
	}

	@Post(":reviewId/like")
	@UseGuards(AuthGuard)
	@ApiBearerAuth()
	@ApiOperation({ summary: "Like a review" })
	@ApiOkResponse({ description: "Review liked" })
	@ApiUnauthorizedResponse({ description: "Not authenticated" })
	async likeReview(
		@Param("reviewId") reviewId: string,
		@Req() req: AuthenticatedRequest,
	): Promise<{ success: boolean }> {
		await this.reviewsService.likeReview(
			req.user.did,
			req.user.session as ATSession,
			reviewId,
		);
		return { success: true };
	}

	@Delete(":reviewId/like")
	@UseGuards(AuthGuard)
	@ApiBearerAuth()
	@ApiOperation({ summary: "Unlike a review" })
	@ApiOkResponse({ description: "Review unliked" })
	@ApiUnauthorizedResponse({ description: "Not authenticated" })
	async unlikeReview(
		@Param("reviewId") reviewId: string,
		@Req() req: AuthenticatedRequest,
	): Promise<{ success: boolean }> {
		await this.reviewsService.unlikeReview(
			req.user.did,
			req.user.session as ATSession,
			reviewId,
		);
		return { success: true };
	}

	@Get(":reviewId/likes")
	@UseGuards(OptionalAuthGuard)
	@ApiOperation({ summary: "Get likes for a review" })
	@ApiOkResponse({
		description: "Likes retrieved",
		type: ReviewLikesResponseDto,
	})
	async getReviewLikes(
		@Param("reviewId") reviewId: string,
		@Req() req: AuthenticatedRequest,
	): Promise<ReviewLikesResponseDto> {
		const result = await this.reviewsService.getReviewLikes(
			reviewId,
			req.user?.did,
		);

		return {
			items: result.items.map((like) => ({
				userDid: like.userDid,
				userHandle: like.userHandle,
				userDisplayName: like.userDisplayName,
				userAvatar: like.userAvatar,
				createdAt: like.createdAt,
			})),
			total: result.total,
			hasLiked: result.hasLiked,
		};
	}

	private toReviewResponse(review: {
		id: string;
		rkey: string;
		title: string;
		markdown: string;
		spoiler: boolean;
		blogDocumentUri: string | null;
		mirrorToBlog: boolean;
		mediaType: string;
		mediaId: string;
		seasonNumber: number;
		episodeNumber: number;
		createdAt: Date;
		updatedAt: Date;
	}): ReviewResponseDto {
		return {
			id: review.id,
			rkey: review.rkey,
			reviewTitle: review.title,
			markdown: review.markdown,
			spoiler: review.spoiler,
			blogDocumentUri: review.blogDocumentUri,
			mirrorToBlog: review.mirrorToBlog,
			mediaType: review.mediaType,
			mediaId: review.mediaId,
			seasonNumber: review.seasonNumber || undefined,
			episodeNumber: review.episodeNumber || undefined,
			createdAt: review.createdAt.toISOString(),
			updatedAt: review.updatedAt.toISOString(),
		};
	}
}